refactor: isCellSelection utility function
