+++ packages/editor/src/core/extensions/table/table/utilities/delete-row.ts
@@ -1,5 +1,7 @@
 import type { Command } from "@tiptap/core";
-import { deleteRow, deleteTable, CellSelection } from "@tiptap/pm/tables";
+import { deleteRow, deleteTable } from "@tiptap/pm/tables";
+// local imports
+import { isCellSelection } from "./helpers";
 
 export const deleteRowOrTable: () => Command =
   () =>
@@ -7,7 +9,7 @@ export const deleteRowOrTable: () => Command =
     const { selection } = state;
 
     // Check if we're in a ProseMirrorTable and have a cell selection
-    if (!(selection instanceof CellSelection)) {
+    if (!isCellSelection(selection)) {
       return false;
     }
