--- packages/editor/src/core/extensions/table/table/utilities/is-cell-selection.ts
@@ -1,5 +1,0 @@
-import { CellSelection } from "@tiptap/pm/tables";
-
-export function isCellSelection(value: unknown): value is CellSelection {
-  return value instanceof CellSelection;
-}
